Update
1. separate delegate
2. fix Unbalanced calls to begin/end appearance transitions…
