+++ JhtNewsChannelSDK/JhtNewsChannel.framework/Headers/JhtNewsChannelItemEditView.h
@@ -15,24 +15,12 @@
 
 /** 频道栏编辑View */
 @interface JhtNewsChannelItemEditView : UIView
-#pragma mark - Block
-/** 返回上级页面的 回调 
- *  modelArr：已添加（上部分）的modelArr
- *  toAddNewModelArr：未添加（下部分）的modelArr
- *  selectIndex：选中的index
- */
-typedef void(^SortViewPop)(NSArray *modelArr, NSArray *nameArray, NSInteger selectIndex, NSArray *toAddNewModelArr);
-
-/** 用于选中通知的block */
-typedef void(^NoticeBlock)();
-
-
 
 #pragma mark - Property
 /** scrollView */
 @property (nonatomic, strong) UIScrollView *sortBigScrollView;
-/** 透明view */
-@property (nonatomic, strong) UIButton *sortBackViewBtn;
+/** 收起排序界面的 透明背景 Button */
+@property (nonatomic, strong) UIButton *sortViewOfPutAwayBtn;
 /** 上级传过来的，装有ChannelModel的数组 */
 @property (nonatomic, strong) NSMutableArray *topBarItemArray;
 /** 上级传过来，装有ChannelModel 待添加的数组 */
@@ -43,7 +31,7 @@ typedef void(^NoticeBlock)();
 @property (nonatomic, strong) UIView *topPartBackView;
 /** @"栏目切换"Label */
 @property (nonatomic, strong) UILabel *tipsLabel;
-/** 排序删除按钮 */
+/** @"排序删除"按钮 */
 @property (nonatomic, strong) UIButton *confirmButton;
 /** 固定位置频道名字的数组 */
 @property (nonatomic, strong) NSArray *notMoveNameArr;
@@ -55,6 +43,15 @@ typedef void(^NoticeBlock)();
 
 
 #pragma mark - Public Method
+/** 返回上级页面的 回调
+ *  modelArr：已添加（上部分）的modelArr
+ *  toAddNewModelArr：未添加（下部分）的modelArr
+ *  selectIndex：选中的index
+ */
+typedef void(^SortViewPopBlock)(NSArray *modelArr, NSArray *nameArray, NSInteger selectIndex, NSArray *toAddNewModelArr);
+/** 用于选中通知的block */
+typedef void(^NoticeBlock)();
+
 /** 初始化
  *  block：用于返回上级页面的 回调
  *  noticeBlock：用于选中通知的block
@@ -65,7 +62,7 @@ typedef void(^NoticeBlock)();
  *  ToAddItemArray：待添加的数组
  * 	itemEditModel：用于排序界面的参数设置
  */
-- (id)initWithTopBarItemArray:(NSArray *)topBarItemArray withToAddItemArray:(NSArray *)toAddItemArray withSelectedName:(NSString *)name withNotMoveNameArray:(NSArray *)notMoveArr isExistDeleteBtn:(BOOL)isExistDelete withNewsChannelItemEditModel:(JhtNewsChannelItemEditParamModel *)itemEditModel withSortBlock:(SortViewPop)block withNoticeBlock:(NoticeBlock)noticeBlock;
+- (id)initWithTopBarItemArray:(NSArray *)topBarItemArray withToAddItemArray:(NSArray *)toAddItemArray withSelectedName:(NSString *)name withNotMoveNameArray:(NSArray *)notMoveArr isExistDeleteBtn:(BOOL)isExistDelete withNewsChannelItemEditModel:(JhtNewsChannelItemEditParamModel *)itemEditModel withSortBlock:(SortViewPopBlock)block withNoticeBlock:(NoticeBlock)noticeBlock;
 
 /** 收起自己 */
 - (void)packUpSelf;
